Relocate types folder
