+++ src/schemaType/ArrayType.ts
@@ -1,5 +1,5 @@
 import type { ForeignKeyDbDefinition } from '../ForeignKeyDbTransformer';
-import type { GenericObject, MvRecord } from '../shared/types';
+import type { GenericObject, MvRecord } from '../types';
 import { ensureArray } from '../utils';
 import BaseScalarArrayType from './BaseScalarArrayType';
 import type BaseScalarType from './BaseScalarType';
