+++ src/schemaType/BaseScalarType.ts
@@ -1,12 +1,6 @@
 import { cloneDeep, set as setIn, toPath } from 'lodash';
 import { InvalidParameterError } from '../errors';
-import type {
-	DecryptFunc,
-	EncryptFunc,
-	GenericObject,
-	MvRecord,
-	SchemaValidator,
-} from '../shared/types';
+import type { DecryptFunc, EncryptFunc, GenericObject, MvRecord, SchemaValidator } from '../types';
 import { getFromMvArray, handleRequiredValidation } from '../utils';
 import BaseSchemaType from './BaseSchemaType';
 import type { SchemaTypeDefinitionBoolean } from './BooleanType';
