+++ src/schemaType/BaseSchemaType.ts
@@ -1,5 +1,5 @@
 import type { ForeignKeyDbDefinition } from '../ForeignKeyDbTransformer';
-import type { GenericObject, MvRecord, Validator } from '../shared/types';
+import type { GenericObject, MvRecord, Validator } from '../types';
 
 export interface SchemaTypeDefinitionBase {
 	path: `${number}` | `${number}.${number}` | `${number}.${number}.${number}` | number;
