+++ src/schemaType/EmbeddedType.ts
@@ -1,7 +1,7 @@
 import { cloneDeep, isPlainObject, set as setIn } from 'lodash';
 import Document from '../Document';
 import type Schema from '../Schema';
-import type { MvRecord } from '../shared/types';
+import type { MvRecord } from '../types';
 import { ensureArray } from '../utils';
 import BaseSchemaType from './BaseSchemaType';
