+++ src/schemaType/ISOCalendarDateTimeType.ts
@@ -1,5 +1,5 @@
 import { TransformDataError } from '../errors';
-import type { GenericObject } from '../shared/types';
+import type { GenericObject } from '../types';
 import { handleTypeValidation } from '../utils';
 import BaseDateType from './BaseDateType';
 import type { ScalarTypeConstructorOptions } from './BaseScalarType';
